Share the mode resolution and ValOps construction
The tool registry needs the same answers createValApiRouter already works
out: which mode we are in, which credential that implies, and which ValOps
implementation to build. Two copies would drift, and the failure would be
quiet - a registry that decides it is in fs mode while the Studio decides it
is in proxy mode reads different content from the same project.

initHandlerOptions moves out of ValRouter verbatim and createValOps is
lifted out of ValServer's constructor block; both call sites now go through
the shared module. No behaviour change.
